(PC-33602)[PRO] fix: move useFunctionOnce hook to commons/hooks directory
